# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Workload Tracker API",
        "version": "0.1.0",
        "description": "OpenAPI schema for Workload Tracker (People, Projects, Assignments, Deliverables, Departments)."
    },
    "paths": {
...
        "/api/roles/{id}/": {
            "get": {
                "operationId": "roles_retrieve",
                "description": "Role ViewSet providing CRUD operations\n- List all roles (paginated)\n- Create new role\n- Retrieve specific role\n- Update role\n- Delete role\n- Bulk list (for autocomplete)",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "A unique integer value identifying this Role.",
                        "required": true
                    }
                ],
                "tags": [
                    "roles"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "put": {
                "operationId": "roles_update",
                "description": "Role ViewSet providing CRUD operations\n- List all roles (paginated)\n- Create new role\n- Retrieve specific role\n- Update role\n- Delete role\n- Bulk list (for autocomplete)",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "A unique integer value identifying this Role.",
                        "required": true
                    }
                ],
                "tags": [
                    "roles"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "patch": {
                "operationId": "roles_partial_update",
                "description": "Role ViewSet providing CRUD operations\n- List all roles (paginated)\n- Create new role\n- Retrieve specific role\n- Update role\n- Delete role\n- Bulk list (for autocomplete)",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "A unique integer value identifying this Role.",
                        "required": true
                    }
                ],
                "tags": [
                    "roles"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedRoleRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedRoleRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedRoleRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "roles_destroy",
                "description": "Override destroy to check for role usage before deletion",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "A unique integer value identifying this Role.",
                        "required": true
                    }
                ],
                "tags": [
                    "roles"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PatchedRoleRequest": {
                "type": "object",
                "description": "Role serializer with camelCase field transformation",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "description": "Role name (e.g., Senior Engineer, Product Manager)",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "description": "Optional description of the role responsibilities"
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Whether this role is currently available for assignment"
                    }
                }
            },
...
            "Role": {
                "type": "object",
                "description": "Role serializer with camelCase field transformation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "description": "Role name (e.g., Senior Engineer, Product Manager)",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "description": "Optional description of the role responsibilities"
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Whether this role is currently available for assignment"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "createdAt",
                    "id",
                    "name",
                    "updatedAt"
                ]
            },
...
            "RoleRequest": {
                "type": "object",
                "description": "Role serializer with camelCase field transformation",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "description": "Role name (e.g., Senior Engineer, Product Manager)",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "description": "Optional description of the role responsibilities"
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Whether this role is currently available for assignment"
                    }
                },
                "required": [
                    "name"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}